bootcfg/: Rename and document packages with their purpose
