+++ bootcfg/cli/doc.go
@@ -1,0 +1,2 @@
+// Package cli provides a command line interface client.
+package cli
