+++ bootcfg/client/doc.go
@@ -1,0 +1,2 @@
+// Package client provides the bootcfg gRPC client.
+package client
