+++ bootcfg/http/doc.go
@@ -1,0 +1,2 @@
+// Package http provides the bootcfg HTTP server
+package http
