+++ bootcfg/http/http.go
@@ -1,4 +1,4 @@
-package api
+package http
 
 import (
 	"net"
