+++ bootcfg/http/ipxe_test.go
@@ -1,4 +1,4 @@
-package api
+package http
 
 import (
 	"net/http"
