+++ bootcfg/http/serialize.go
@@ -1,4 +1,4 @@
-package api
+package http
 
 import (
 	"encoding/json"
