+++ bootcfg/rpc/doc.go
@@ -1,0 +1,2 @@
+// Package rpc provides the bootcfg gRPC server
+package rpc
