+++ bootcfg/server/doc.go
@@ -1,0 +1,2 @@
+// Package server is a bootcfg library package for implementing servers.
+package server
