+++ bootcfg/sign/doc.go
@@ -1,0 +1,2 @@
+// Package sign adds signatures to bootcfg responses.
+package sign
