+++ bootcfg/storage/doc.go
@@ -1,0 +1,2 @@
+// Package storage defines bootcfg's storage and object types.
+package storage
