+++ cmd/bootcmd/main.go
@@ -3,5 +3,5 @@ package main
 import "github.com/coreos/coreos-baremetal/bootcfg/cli"
 
 func main() {
-	cmd.Execute()
+	cli.Execute()
 }
